Email alerts uses presenter for formatting payload as well as email body
